change preprocessing to extraction and modality to preprocessing
