+++ clinicadl/commandline/pipelines/generate/shepplogan/cli.py
@@ -6,6 +6,7 @@ import pandas as pd
 import torch
 from joblib import Parallel, delayed
 
+from clinicadl.caps_dataset.extraction.utils import write_preprocessing
 from clinicadl.commandline import arguments
 from clinicadl.commandline.modules_options import data, dataloader
 from clinicadl.commandline.pipelines.generate.shepplogan import options as shepplogan
@@ -14,7 +15,6 @@ from clinicadl.generate.generate_utils import (
     generate_shepplogan_phantom,
     write_missing_mods,
 )
-from clinicadl.preprocessing.preprocessing import write_preprocessing
 from clinicadl.utils.maps_manager.iotools import check_and_clean, commandline_to_json
 
 logger = getLogger("clinicadl.generate.shepplogan")
