+++ clinicadl/commandline/pipelines/prepare_data/prepare_data_from_bids_cli.py
@@ -5,7 +5,7 @@ from clinicadl.commandline import arguments
 from clinicadl.commandline.modules_options import (
     data,
     dataloader,
-    modality,
+    extraction,
     preprocessing,
 )
 from clinicadl.prepare_data.prepare_data import DeepLearningPrepareData
@@ -17,11 +17,11 @@ from clinicadl.utils.enum import ExtractionMethod
 @arguments.caps_directory
 @arguments.modality_bids
 @dataloader.n_proc
-@preprocessing.extract_json
-@preprocessing.use_uncropped_image
-@modality.tracer
-@modality.suvr_reference_region
-@modality.custom_suffix
+@extraction.extract_json
+@extraction.use_uncropped_image
+@preprocessing.tracer
+@preprocessing.suvr_reference_region
+@preprocessing.custom_suffix
 @data.participants_tsv
 def image_bids_cli(kwargs):
     image_config = CapsDatasetConfig.from_preprocessing_and_extraction_method(
@@ -38,15 +38,15 @@ def image_bids_cli(kwargs):
 @arguments.caps_directory
 @arguments.modality_bids
 @dataloader.n_proc
-@preprocessing.save_features
+@extraction.save_features
 @data.participants_tsv
-@preprocessing.extract_json
-@preprocessing.use_uncropped_image
-@preprocessing.patch_size
-@preprocessing.stride_size
-@modality.tracer
-@modality.suvr_reference_region
-@modality.custom_suffix
+@extraction.extract_json
+@extraction.use_uncropped_image
+@extraction.patch_size
+@extraction.stride_size
+@preprocessing.tracer
+@preprocessing.suvr_reference_region
+@preprocessing.custom_suffix
 def patch_bids_cli(kwargs):
     """Extract patch from nifti images.
     CAPS_DIRECTORY is the CAPS folder where nifti images are stored and tensor will be saved.
@@ -66,16 +66,16 @@ def patch_bids_cli(kwargs):
 @arguments.caps_directory
 @arguments.modality_bids
 @dataloader.n_proc
-@preprocessing.save_features
+@extraction.save_features
 @data.participants_tsv
-@preprocessing.extract_json
-@preprocessing.use_uncropped_image
-@preprocessing.slice_direction
-@preprocessing.slice_mode
-@preprocessing.discarded_slices
-@modality.tracer
-@modality.suvr_reference_region
-@modality.custom_suffix
+@extraction.extract_json
+@extraction.use_uncropped_image
+@extraction.slice_direction
+@extraction.slice_mode
+@extraction.discarded_slices
+@preprocessing.tracer
+@preprocessing.suvr_reference_region
+@preprocessing.custom_suffix
 def slice_bids_cli(kwargs):
     """Extract slice from nifti images.
     CAPS_DIRECTORY is the CAPS folder where nifti images are stored and tensor will be saved.
@@ -96,17 +96,17 @@ def slice_bids_cli(kwargs):
 @arguments.caps_directory
 @arguments.modality_bids
 @dataloader.n_proc
-@preprocessing.save_features
+@extraction.save_features
 @data.participants_tsv
-@preprocessing.extract_json
-@preprocessing.use_uncropped_image
-@preprocessing.roi_custom_mask_pattern
-@preprocessing.roi_custom_template
-@preprocessing.roi_list
-@preprocessing.roi_uncrop_output
-@modality.tracer
-@modality.suvr_reference_region
-@modality.custom_suffix
+@extraction.extract_json
+@extraction.use_uncropped_image
+@extraction.roi_custom_mask_pattern
+@extraction.roi_custom_template
+@extraction.roi_list
+@extraction.roi_uncrop_output
+@preprocessing.tracer
+@preprocessing.suvr_reference_region
+@preprocessing.custom_suffix
 def roi_bids_cli(kwargs):
     """Extract roi from nifti images.
