+++ clinicadl/prepare_data/prepare_data.py
@@ -7,14 +7,14 @@ from torch import save as save_tensor
 
 from clinicadl.caps_dataset.caps_dataset_config import CapsDatasetConfig
 from clinicadl.caps_dataset.caps_dataset_utils import compute_folder_and_file_type
-from clinicadl.preprocessing.config import (
-    PreprocessingConfig,
-    PreprocessingImageConfig,
-    PreprocessingPatchConfig,
-    PreprocessingROIConfig,
-    PreprocessingSliceConfig,
+from clinicadl.caps_dataset.extraction.config import (
+    ExtractionConfig,
+    ExtractionImageConfig,
+    ExtractionPatchConfig,
+    ExtractionROIConfig,
+    ExtractionSliceConfig,
 )
-from clinicadl.preprocessing.preprocessing import write_preprocessing
+from clinicadl.caps_dataset.extraction.utils import write_preprocessing
 from clinicadl.utils.clinica_utils import (
     check_caps_folder,
     clinicadl_file_reader,
@@ -49,16 +49,16 @@ def DeepLearningPrepareData(
         input_directory, config.data.data_tsv, is_bids_dir, False, None
     )
 
-    if config.preprocessing.save_features:
+    if config.extraction.save_features:
         logger.info(
-            f"{config.preprocessing.extract_method.value}s will be extracted in Pytorch tensor from {len(sessions)} images."
+            f"{config.extraction.extract_method.value}s will be extracted in Pytorch tensor from {len(sessions)} images."
         )
     else:
         logger.info(
             f"Images will be extracted in Pytorch tensor from {len(sessions)} images."
         )
         logger.info(
-            f"Information for {config.preprocessing.extract_method.value} will be saved in output JSON file and will be used "
+            f"Information for {config.extraction.extract_method.value} will be saved in output JSON file and will be used "
             f"during training for on-the-fly extraction."
         )
     logger.debug(f"List of subjects: \n{subjects}.")
@@ -95,8 +95,8 @@ def DeepLearningPrepareData(
             logger.debug(f"Output tensor saved at {output_file}")
 
     if (
-        config.preprocessing.extract_method == ExtractionMethod.IMAGE
-        or not config.preprocessing.save_features
+        config.extraction.extract_method == ExtractionMethod.IMAGE
+        or not config.extraction.save_features
     ):
 
         def prepare_image(file):
@@ -113,22 +113,22 @@ def DeepLearningPrepareData(
             delayed(prepare_image)(file) for file in input_files
         )
 
-    elif config.preprocessing.save_features:
-        if config.preprocessing.extract_method == ExtractionMethod.SLICE:
-            assert isinstance(config.preprocessing, PreprocessingSliceConfig)
+    elif config.extraction.save_features:
+        if config.extraction.extract_method == ExtractionMethod.SLICE:
+            assert isinstance(config.extraction, ExtractionSliceConfig)
 
             def prepare_slice(file):
                 from .prepare_data_utils import extract_slices
 
-                assert isinstance(config.preprocessing, PreprocessingSliceConfig)
+                assert isinstance(config.extraction, ExtractionSliceConfig)
                 logger.debug(f"  Processing of {file}.")
                 container = container_from_filename(file)
                 subfolder = "slice_based"
                 output_mode = extract_slices(
                     Path(file),
-                    slice_direction=config.preprocessing.slice_direction,
-                    slice_mode=config.preprocessing.slice_mode,
-                    discarded_slices=config.preprocessing.discarded_slices,
+                    slice_direction=config.extraction.slice_direction,
+                    slice_mode=config.extraction.slice_mode,
+                    discarded_slices=config.extraction.discarded_slices,
                 )
                 logger.debug(f"    {len(output_mode)} slices extracted.")
                 write_output_imgs(output_mode, container, subfolder)
@@ -137,20 +137,20 @@ def DeepLearningPrepareData(
                 delayed(prepare_slice)(file) for file in input_files
             )
 
-        elif config.preprocessing.extract_method == ExtractionMethod.PATCH:
-            assert isinstance(config.preprocessing, PreprocessingPatchConfig)
+        elif config.extraction.extract_method == ExtractionMethod.PATCH:
+            assert isinstance(config.extraction, ExtractionPatchConfig)
 
             def prepare_patch(file):
                 from .prepare_data_utils import extract_patches
 
-                assert isinstance(config.preprocessing, PreprocessingPatchConfig)
+                assert isinstance(config.extraction, ExtractionPatchConfig)
                 logger.debug(f"  Processing of {file}.")
                 container = container_from_filename(file)
                 subfolder = "patch_based"
                 output_mode = extract_patches(
                     Path(file),
-                    patch_size=config.preprocessing.patch_size,
-                    stride_size=config.preprocessing.stride_size,
+                    patch_size=config.extraction.patch_size,
+                    stride_size=config.extraction.stride_size,
                 )
                 logger.debug(f"    {len(output_mode)} patches extracted.")
                 write_output_imgs(output_mode, container, subfolder)
@@ -159,23 +159,23 @@ def DeepLearningPrepareData(
                 delayed(prepare_patch)(file) for file in input_files
             )
 
-        elif config.preprocessing.extract_method == ExtractionMethod.ROI:
-            assert isinstance(config.preprocessing, PreprocessingROIConfig)
+        elif config.extraction.extract_method == ExtractionMethod.ROI:
+            assert isinstance(config.extraction, ExtractionROIConfig)
 
             def prepare_roi(file):
                 from .prepare_data_utils import extract_roi
 
-                assert isinstance(config.preprocessing, PreprocessingROIConfig)
+                assert isinstance(config.extraction, ExtractionROIConfig)
                 logger.debug(f"  Processing of {file}.")
                 container = container_from_filename(file)
                 subfolder = "roi_based"
                 if config.preprocessing == Preprocessing.CUSTOM:
-                    if not config.preprocessing.roi_custom_template:
+                    if not config.extraction.roi_custom_template:
                         raise ClinicaDLArgumentError(
                             "A custom template must be defined when the modality is set to custom."
                         )
-                    roi_template = config.preprocessing.roi_custom_template
-                    roi_mask_pattern = config.preprocessing.roi_custom_mask_pattern
+                    roi_template = config.extraction.roi_custom_template
+                    roi_mask_pattern = config.extraction.roi_custom_mask_pattern
                 else:
                     if config.preprocessing.preprocessing == Preprocessing.T1_LINEAR:
                         roi_template = Template.T1_LINEAR
@@ -191,25 +191,25 @@ def DeepLearningPrepareData(
 
                 masks_location = input_directory / "masks" / f"tpl-{roi_template}"
 
-                if len(config.preprocessing.roi_list) == 0:
+                if len(config.extraction.roi_list) == 0:
                     raise ClinicaDLArgumentError(
                         "A list of regions of interest must be given."
                     )
                 else:
                     check_mask_list(
                         masks_location,
-                        config.preprocessing.roi_list,
+                        config.extraction.roi_list,
                         roi_mask_pattern,
-                        config.preprocessing.use_uncropped_image,
+                        config.extraction.use_uncropped_image,
                     )
 
                 output_mode = extract_roi(
                     Path(file),
                     masks_location=masks_location,
                     mask_pattern=roi_mask_pattern,
-                    cropped_input=not config.preprocessing.use_uncropped_image,
-                    roi_names=config.preprocessing.roi_list,
-                    uncrop_output=config.preprocessing.roi_uncrop_output,
+                    cropped_input=not config.extraction.use_uncropped_image,
+                    roi_names=config.extraction.roi_list,
+                    uncrop_output=config.extraction.roi_uncrop_output,
                 )
                 logger.debug("ROI extracted.")
                 write_output_imgs(output_mode, container, subfolder)
@@ -220,11 +220,11 @@ def DeepLearningPrepareData(
 
     else:
         raise NotImplementedError(
-            f"Extraction is not implemented for mode {config.preprocessing.extract_method.value}."
+            f"Extraction is not implemented for mode {config.extraction.extract_method.value}."
         )
 
     # Save parameters dictionary
     preprocessing_json_path = write_preprocessing(
-        config.preprocessing.model_dump(), config.data.caps_directory
+        config.extraction.model_dump(), config.data.caps_directory
     )
     logger.info(f"Preprocessing JSON saved at {preprocessing_json_path}.")
